nptel w6assignment4.java
Due on 2024-09-05, 23:59 IST
Execution of two or more threads occurs in a random order. The keyword 'synchronized' in Java is used to control the execution of thread in a strict sequence. In the following, the program is expected to print some numbers. Do the necessary use of 'synchronized' keyword, so that, the program prints the output in the following order:  
-----------------OUTPUT------------------- (this line should not to be printed)
5
10
15
20
25
100
200
300
400
500
-------------------------------------------------(this line should not to be printed)