nptel w10assignment4.java
A string "s1" is already initialized. You have to read the index "n".
Complete the code segment to catch the exception in the following, if any.
On the occurrence of such an exception, your program should print “exception occur”.
If there is no such exception, your program should replace the char ‘a’ at the index value "n" of the "s1",
then it will print the modified string.